revert: "feat: add platform role to user"

+++ packages/server/api/src/app/ee/flow-template/platform-flow-template.module.ts
@@ -1,15 +1,17 @@
 import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox'
 import { Static, Type } from '@sinclair/typebox'
 import { StatusCodes } from 'http-status-codes'
-import { platformMustBeOwnedByCurrentUser } from '../authentication/ee-authorization'
+import { platformService } from '../../platform/platform.service'
 import { flowTemplateService } from './flow-template.service'
 import { CreateFlowTemplateRequest } from '@activepieces/ee-shared'
 import { system, SystemProp } from '@activepieces/server-shared'
 import {
     ActivepiecesError,
     ALL_PRINCIPAL_TYPES,
+    assertNotNullOrUndefined,
     ErrorCode,
     ListFlowTemplatesRequest,
+    Principal,
     PrincipalType,
     TemplateType,
 } from '@activepieces/shared'
@@ -53,7 +55,7 @@ const flowTemplateController: FastifyPluginAsyncTypebox = async (fastify) => {
         },
         async (request) => {
             const platformId =
-                request.principal.type === PrincipalType.UNKNOWN ? system.getOrThrow(SystemProp.CLOUD_PLATFORM_ID) : request.principal.platform.id
+        request.principal.type === PrincipalType.UNKNOWN ? system.getOrThrow(SystemProp.CLOUD_PLATFORM_ID) : request.principal.platform.id
             return flowTemplateService.list(platformId, request.query)
         },
     )
@@ -68,10 +70,13 @@ const flowTemplateController: FastifyPluginAsyncTypebox = async (fastify) => {
                 body: CreateFlowTemplateRequest,
             },
         },
-        async (request, reply) => {
+        async (request) => {
             const { type } = request.body
             if (type === TemplateType.PLATFORM) {
-                await platformMustBeOwnedByCurrentUser.call(fastify, request, reply)
+                await assertUserIsPlatformOwner({
+                    platformId: request.principal.platform.id,
+                    userId: request.principal.id,
+                })
             }
             return flowTemplateService.upsert(
                 request.principal.platform.id,
@@ -91,20 +96,11 @@ const flowTemplateController: FastifyPluginAsyncTypebox = async (fastify) => {
             },
         },
         async (request, reply) => {
-            const template = await flowTemplateService.getOrThrow(request.params.id)
-            switch (template.type) {
-                case TemplateType.PLATFORM:
-                    await platformMustBeOwnedByCurrentUser.call(fastify, request, reply)
-                    break
-                case TemplateType.PROJECT:
-                    if (template.projectId !== request.principal.projectId) {
-                        throw new ActivepiecesError({
-                            code: ErrorCode.AUTHORIZATION,
-                            params: {},
-                        })
-                    }
-                    break
-            }
+            await assertUserCanDeleteTemplate({
+                templateId: request.params.id,
+                userId: request.principal.id,
+                principal: request.principal,
+            })
             await flowTemplateService.delete({
                 id: request.params.id,
             })
@@ -113,3 +109,50 @@ const flowTemplateController: FastifyPluginAsyncTypebox = async (fastify) => {
     )
 }
 
+async function assertUserCanDeleteTemplate({
+    templateId,
+    userId,
+    principal,
+}: {
+    templateId: string
+    userId: string
+    principal: Principal
+}): Promise<void> {
+    const template = await flowTemplateService.getOrThrow(templateId)
+    switch (template.type) {
+        case TemplateType.PLATFORM:
+            await assertUserIsPlatformOwner({
+                platformId: template.platformId,
+                userId,
+            })
+            break
+        case TemplateType.PROJECT:
+            if (template.projectId !== principal.projectId) {
+                throw new ActivepiecesError({
+                    code: ErrorCode.AUTHORIZATION,
+                    params: {},
+                })
+            }
+            break
+    }
+}
+
+async function assertUserIsPlatformOwner({
+    platformId,
+    userId,
+}: {
+    platformId?: string
+    userId: string
+}): Promise<void> {
+    assertNotNullOrUndefined(platformId, 'platformId')
+    const userOwner = await platformService.checkUserIsOwner({
+        platformId,
+        userId,
+    })
+    if (!userOwner) {
+        throw new ActivepiecesError({
+            code: ErrorCode.AUTHORIZATION,
+            params: {},
+        })
+    }
+}
